requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
semantic_kernel[google_palm]==0.9.0.beta1
azure-identity
azure-cosmos